[WIP] Use shared informer and informer cache
This refactors everything to rely on the inforer cache and a work queue
provided with the k8s client.

The work queue ensures that only one operation happens at a time on a
given key, supports requeing with tracking retries, and backoff on
retries.

This mostly removes the need for the resource manager, however it's
still used in some places (e.g. to get secrets) that should be
refactored separately.

This change fixes concurrency issues when operating on a pod by ensuring
that only one operation occurs at a time on a pod and ensures the worker
must get the most up to date version of a pod.